MM-54201 Refactor current mention parsing into MentionParserStandard (#24936)
* MM-54201 Refactor current mention parsing into MentionParserStandard

* Fix vet
