+++ server/channels/app/notification.go
@@ -10,8 +10,6 @@ import (
 	"sort"
 	"strings"
 	"sync"
-	"unicode"
-	"unicode/utf8"
 
 	"github.com/pkg/errors"
 
@@ -1048,25 +1046,36 @@ func splitAtFinal(items []string) (preliminary []string, final string) {
 // Given a message and a map mapping mention keywords to the users who use them, returns a map of mentioned
 // users and a slice of potential mention users not in the channel and whether or not @here was mentioned.
 func getExplicitMentions(post *model.Post, keywords map[string][]string, groups map[string]*model.Group) *MentionResults {
-	ret := &MentionResults{}
+	parser := makeStandardMentionParser(keywords, groups)
 
 	buf := ""
 	mentionsEnabledFields := getMentionsEnabledFields(post)
 	for _, message := range mentionsEnabledFields {
+		// Parse the text as Markdown, combining adjacent Text nodes into a single string for processing
 		markdown.Inspect(message, func(node any) bool {
 			text, ok := node.(*markdown.Text)
 			if !ok {
-				processText(ret, buf, keywords, groups)
+				// This node isn't a string so process any accumulated text in the buffer
+				if buf != "" {
+					parser.ProcessText(buf)
+				}
+
 				buf = ""
 				return true
 			}
+
+			// This node is a string, so add it to buf and continue onto the next node to see if it's more text
 			buf += text.Text
 			return false
 		})
 	}
-	processText(ret, buf, keywords, groups)
 
-	return ret
+	// Process any left over text
+	if buf != "" {
+		parser.ProcessText(buf)
+	}
+
+	return parser.Results()
 }
 
 // Given a post returns the values of the fields in which mentions are possible.
@@ -1303,155 +1312,6 @@ func (n *PostNotification) GetSenderName(userNameFormat string, overridesAllowed
 	}
 
 	return n.Sender.GetDisplayNameWithPrefix(userNameFormat, "@")
-}
-
-// checkForMention checks if there is a mention to a specific user or to the keywords here / channel / all
-func checkForMention(m *MentionResults, word string, keywords map[string][]string, groups map[string]*model.Group) bool {
-	var mentionType MentionType
-
-	switch strings.ToLower(word) {
-	case "@here":
-		m.HereMentioned = true
-		mentionType = ChannelMention
-	case "@channel":
-		m.ChannelMentioned = true
-		mentionType = ChannelMention
-	case "@all":
-		m.AllMentioned = true
-		mentionType = ChannelMention
-	default:
-		mentionType = KeywordMention
-	}
-
-	checkForGroupMention(m, word, groups)
-
-	if ids, match := keywords[strings.ToLower(word)]; match {
-		m.addMentions(ids, mentionType)
-		return true
-	}
-
-	// Case-sensitive check for first name
-	if ids, match := keywords[word]; match {
-		m.addMentions(ids, mentionType)
-		return true
-	}
-
-	return false
-}
-
-func checkForGroupMention(m *MentionResults, word string, groups map[string]*model.Group) bool {
-	if strings.HasPrefix(word, "@") {
-		word = word[1:]
-	} else {
-		// Only allow group mentions when mentioned directly with @group-name
-		return false
-	}
-
-	group, groupFound := groups[word]
-	if !groupFound {
-		group = groups[strings.ToLower(word)]
-	}
-
-	if group == nil {
-		return false
-	}
-
-	if m.GroupMentions == nil {
-		m.GroupMentions = make(map[string]*model.Group)
-	}
-
-	if group.Name != nil {
-		m.GroupMentions[*group.Name] = group
-	}
-
-	return true
-}
-
-// isKeywordMultibyte checks if a word containing a multibyte character contains a multibyte keyword
-func isKeywordMultibyte(keywords map[string][]string, word string) ([]string, bool) {
-	ids := []string{}
-	match := false
-	var multibyteKeywords []string
-	for keyword := range keywords {
-		if len(keyword) != utf8.RuneCountInString(keyword) {
-			multibyteKeywords = append(multibyteKeywords, keyword)
-		}
-	}
-
-	if len(word) != utf8.RuneCountInString(word) {
-		for _, key := range multibyteKeywords {
-			if strings.Contains(word, key) {
-				ids, match = keywords[key]
-			}
-		}
-	}
-	return ids, match
-}
-
-// Processes text to filter mentioned users and other potential mentions
-func processText(m *MentionResults, text string, keywords map[string][]string, groups map[string]*model.Group) {
-	systemMentions := map[string]bool{"@here": true, "@channel": true, "@all": true}
-
-	for _, word := range strings.FieldsFunc(text, func(c rune) bool {
-		// Split on any whitespace or punctuation that can't be part of an at mention or emoji pattern
-		return !(c == ':' || c == '.' || c == '-' || c == '_' || c == '@' || unicode.IsLetter(c) || unicode.IsNumber(c))
-	}) {
-		// skip word with format ':word:' with an assumption that it is an emoji format only
-		if word[0] == ':' && word[len(word)-1] == ':' {
-			continue
-		}
-
-		word = strings.TrimLeft(word, ":.-_")
-
-		if checkForMention(m, word, keywords, groups) {
-			continue
-		}
-
-		foundWithoutSuffix := false
-		wordWithoutSuffix := word
-
-		for wordWithoutSuffix != "" && strings.LastIndexAny(wordWithoutSuffix, ".-:_") == (len(wordWithoutSuffix)-1) {
-			wordWithoutSuffix = wordWithoutSuffix[0 : len(wordWithoutSuffix)-1]
-
-			if checkForMention(m, wordWithoutSuffix, keywords, groups) {
-				foundWithoutSuffix = true
-				break
-			}
-		}
-
-		if foundWithoutSuffix {
-			continue
-		}
-
-		if _, ok := systemMentions[word]; !ok && strings.HasPrefix(word, "@") {
-			// No need to bother about unicode as we are looking for ASCII characters.
-			last := word[len(word)-1]
-			switch last {
-			// If the word is possibly at the end of a sentence, remove that character.
-			case '.', '-', ':':
-				word = word[:len(word)-1]
-			}
-			m.OtherPotentialMentions = append(m.OtherPotentialMentions, word[1:])
-		} else if strings.ContainsAny(word, ".-:") {
-			// This word contains a character that may be the end of a sentence, so split further
-			splitWords := strings.FieldsFunc(word, func(c rune) bool {
-				return c == '.' || c == '-' || c == ':'
-			})
-
-			for _, splitWord := range splitWords {
-				if checkForMention(m, splitWord, keywords, groups) {
-					continue
-				}
-				if _, ok := systemMentions[splitWord]; !ok && strings.HasPrefix(splitWord, "@") {
-					m.OtherPotentialMentions = append(m.OtherPotentialMentions, splitWord[1:])
-				}
-			}
-		}
-
-		if ids, match := isKeywordMultibyte(keywords, word); match {
-			m.addMentions(ids, KeywordMention)
-		}
-	}
 }
 
 func (a *App) GetNotificationNameFormat(user *model.User) string {
